+++ android/Staccato_AN/app/src/main/java/com/woowacourse/staccato/data/dto/moment/VisitCreationRequest.kt
@@ -1,14 +1,14 @@
-package com.woowacourse.staccato.data.dto.visit
+package com.woowacourse.staccato.data.dto.moment
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class VisitResponse(
-    @SerialName("visitId") val visitId: Long,
+data class VisitCreationRequest(
+    @SerialName("memoryId") val memoryId: Long,
     @SerialName("placeName") val placeName: String,
-    @SerialName("visitImageUrls") val visitImageUrls: List<String>,
+    @SerialName("latitude") val latitude: String,
+    @SerialName("longitude") val longitude: String,
     @SerialName("address") val address: String,
     @SerialName("visitedAt") val visitedAt: String,
-    @SerialName("visitLogs") val visitLogs: List<VisitLogDto>,
 )
