+++ android/Staccato_AN/app/src/main/java/com/woowacourse/staccato/data/dto/moment/VisitCreationResponse.kt
@@ -1,4 +1,4 @@
-package com.woowacourse.staccato.data.dto.visit
+package com.woowacourse.staccato.data.dto.moment
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
